Remove gratuitous FunctionLoader interface
